refactor(config): extract the Field-default guard to tools-lint and record the size shrinks (#14070)
