+++ repo_tests/python_file_size_ratchet_baseline.py
@@ -321,10 +321,10 @@ RATCHET_BASELINE: dict[str, int] = {
     "autobot-backend/secure_sandbox_executor.py": 1019,
     "autobot-backend/security/command_patterns.py": 609,
     "autobot-backend/security/domain_security.py": 725,
-    "autobot-backend/security/enterprise/compliance_manager.py": 922,
-    "autobot-backend/security/enterprise/security_policy_manager.py": 968,
-    "autobot-backend/security/enterprise/sso_integration.py": 928,
-    "autobot-backend/security/enterprise/threat_detection/engine.py": 731,
+    "autobot-backend/security/enterprise/compliance_manager.py": 911,
+    "autobot-backend/security/enterprise/security_policy_manager.py": 957,
+    "autobot-backend/security/enterprise/sso_integration.py": 917,
+    "autobot-backend/security/enterprise/threat_detection/engine.py": 720,
     "autobot-backend/security/input_validator.py": 607,
     "autobot-backend/security/prompt_injection_detector.py": 747,
     "autobot-backend/security/security_edge_cases_test.py": 638,
@@ -332,7 +332,6 @@ RATCHET_BASELINE: dict[str, int] = {
     "autobot-backend/security/threat_detection_refactor_test.py": 880,
     "autobot-backend/security/threat_intelligence.py": 771,
     "autobot-backend/security_layer.py": 767,
-    "autobot-backend/security_layer_test.py": 604,
     "autobot-backend/services/agent_analytics.py": 686,
     "autobot-backend/services/agent_terminal/service.py": 1015,
     "autobot-backend/services/ai_stack_client.py": 790,
@@ -528,8 +527,8 @@ RATCHET_BASELINE: dict[str, int] = {
     "autobot_shared/redis_client.py": 679,
     "autobot_shared/redis_management/connection_manager.py": 1519,
     "autobot_shared/security/ssrf_guard_test.py": 633,
-    "autobot_shared/ssot_config.py": 3299,
-    "autobot_shared/ssot_config_test.py": 784,
+    "autobot_shared/ssot_config.py": 3297,
+    "autobot_shared/ssot_config_test.py": 782,
     "autobot_shared/ssot_constants.py": 958,
     "autobot_shared/user_management/team_service.py": 725,
     "pipeline-scripts/check-pre-commit-hook-pr_test.py": 751,
